% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{add two numbers}
\usage{
f(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
sum of two numbers
}
\description{
Add up two numbers (Description)
}
\details{
We want to add up two numbers, blalala... (Details)
}
\examples{
f(1,2)
}
\author{
Caleb
}
