% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Substract two numbers}
\usage{
g(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
x - y
}
\description{
Substract two numbers (Description)
}
\details{
We want to substract two numbers, blalala... (Details)
}
\examples{
g(2, 1)
}
\author{
Seven
}
