% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{Very complex operation of two numbers}
\usage{
h(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
(x + y)(x - y)
}
\description{
Complex operations on two numbers (Description)
}
\details{
We want to do some complex operations on two numbers, blalala... (Details)
}
\examples{
h(3, 2)
}
\author{
Seven
}
