% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{all}
\alias{all}
\title{return all}
\usage{
all(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
A list of f, g, and h.
\item{f}{Results for adding}
\item{g}{Results for subtracting}
\item{h}{Complex Result}
}
\description{
Return f,g,h (Description)
}
\details{
We want to return f,g,h , blalala... (Details)
}
\examples{
all(3,2)
}
\author{
Caleb
}
